// Agileo-FrontEnd/src/interfaces/ireception.ts
export interface IReception {
  id?: number;
  affaireId?: string;
  userId?: number;
  dateReception?: string;
  referenceBl?: string;
  dateBl?: string;
  refFournisseur?: string;
  nomFournisseur?: string;
  idAgelio?: number;
  blDivalto?: number;
  statut?: string;
  affaireCode?: string;
  userLogin?: string;
  createdBy?: string;
  createdDate?: string;
  updatedBy?: string;
  updatedDate?: string;
  affaireLibelle?: string;
  createurNom?: string;
  commandeCode?: number;
}